requirements.txt
flask
streamlit
azure-openai
pyyaml
python-dotenv
pydantic